# pkg/apiclient/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: v0.0.0-dev
servers:
- url: http://localhost:3986/
security:
- Bearer: []
paths:
...
  /target:
    get:
      description: List targets
      operationId: ListTargets
      parameters:
      - description: Show target config options
        in: query
        name: showOptions
        required: true
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TargetDTO'
                type: array
          description: OK
      summary: List targets
      tags:
      - target
    post:
      description: Create a target
      operationId: CreateTarget
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateTargetDTO'
        description: Create target
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
          description: OK
      summary: Create a target
      tags:
      - target
      x-codegen-request-body-name: target
  /target-config:
    get:
      description: List target configs
      operationId: ListTargetConfigs
      parameters:
      - description: Show target config options
        in: query
        name: showOptions
        required: true
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TargetConfig'
                type: array
          description: OK
      summary: List target configs
      tags:
      - target-config
    put:
      description: Add a target config
      operationId: AddTargetConfig
      parameters:
      - description: Show target config options
        in: query
        name: showOptions
        required: true
        schema:
          type: boolean
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/AddTargetConfigDTO'
        description: Target config to add
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TargetConfig'
          description: OK
      summary: Add a target config
      tags:
      - target-config
      x-codegen-request-body-name: targetConfig
...
  /target/{targetId}:
    delete:
      description: Remove target
      operationId: RemoveTarget
      parameters:
      - description: Target ID
        in: path
        name: targetId
        required: true
        schema:
          type: string
      - description: Force
        in: query
        name: force
        schema:
          type: boolean
      responses:
        "200":
          content: {}
          description: OK
      summary: Remove target
      tags:
      - target
    get:
      description: Get target info
      operationId: GetTarget
      parameters:
      - description: Target ID or Name
        in: path
        name: targetId
        required: true
        schema:
          type: string
      - description: Show target config options
        in: query
        name: showOptions
        required: true
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetDTO'
          description: OK
      summary: Get target info
      tags:
      - target
...
components:
  schemas:
    AddTargetConfigDTO:
      example:
        name: name
        options: options
        providerInfo:
          agentlessTarget: true
          name: name
          runnerId: runnerId
          label: label
          version: version
      properties:
        name:
          type: string
        options:
          type: string
        providerInfo:
          $ref: '#/components/schemas/TargetProviderInfo'
      required:
      - name
      - options
      - providerInfo
      type: object
...
    BuildConfig:
      example:
        cachedBuild:
          image: image
          user: user
        devcontainer:
          filePath: filePath
      properties:
        cachedBuild:
          $ref: '#/components/schemas/CachedBuild'
        devcontainer:
          $ref: '#/components/schemas/DevcontainerConfig'
      type: object
...
    CachedBuild:
      example:
        image: image
        user: user
      properties:
        image:
          type: string
        user:
          type: string
      required:
      - image
      - user
      type: object
    CloneTarget:
      enum:
      - branch
      - commit
      type: string
      x-enum-varnames:
      - CloneTargetBranch
      - CloneTargetCommit
...
    CreateTargetDTO:
      example:
        targetConfigName: targetConfigName
        name: name
        id: id
      properties:
        id:
          type: string
        name:
          type: string
        targetConfigName:
          type: string
      required:
      - id
      - name
      - targetConfigName
      type: object
...
    DevcontainerConfig:
      example:
        filePath: filePath
      properties:
        filePath:
          type: string
      required:
      - filePath
      type: object
...
    FileStatus:
      example:
        extra: extra
        name: name
        staging: null
        worktree: null
      properties:
        extra:
          type: string
        name:
          type: string
        staging:
          $ref: '#/components/schemas/Status'
        worktree:
          $ref: '#/components/schemas/Status'
      required:
      - extra
      - name
      - staging
      - worktree
      type: object
...
    GitRepository:
      example:
        owner: owner
        path: path
        name: name
        id: id
        source: source
        prNumber: 0
        branch: branch
        cloneTarget: null
        sha: sha
        url: url
      properties:
        branch:
          type: string
        cloneTarget:
          $ref: '#/components/schemas/CloneTarget'
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        path:
          type: string
        prNumber:
          type: integer
        sha:
          type: string
        source:
          type: string
        url:
          type: string
      required:
      - branch
      - id
      - name
      - owner
      - sha
      - source
      - url
      type: object
    GitStatus:
      example:
        behind: 1
        fileStatus:
        - extra: extra
          name: name
          staging: null
          worktree: null
        - extra: extra
          name: name
          staging: null
          worktree: null
        ahead: 6
        branchPublished: true
        currentBranch: currentBranch
      properties:
        ahead:
          type: integer
        behind:
          type: integer
        branchPublished:
          type: boolean
        currentBranch:
          type: string
        fileStatus:
          items:
            $ref: '#/components/schemas/FileStatus'
          type: array
      required:
      - currentBranch
      - fileStatus
      type: object
...
    Job:
      example:
        createdAt: createdAt
        resourceId: resourceId
        action: null
        id: id
        state: null
        error: error
        resourceType: null
        updatedAt: updatedAt
      properties:
        action:
          $ref: '#/components/schemas/models.JobAction'
        createdAt:
          type: string
        error:
          type: string
        id:
          type: string
        resourceId:
          type: string
        resourceType:
          $ref: '#/components/schemas/ResourceType'
        state:
          $ref: '#/components/schemas/JobState'
        updatedAt:
          type: string
      required:
      - action
      - createdAt
      - id
      - resourceId
      - resourceType
      - state
      - updatedAt
      type: object
    JobState:
      enum:
      - pending
      - running
      - error
      - success
      type: string
      x-enum-varnames:
      - JobStatePending
      - JobStateRunning
      - JobStateError
      - JobStateSuccess
...
    ResourceState:
      example:
        name: null
        error: error
        updatedAt: updatedAt
      properties:
        error:
          type: string
        name:
          $ref: '#/components/schemas/models.ResourceStateName'
        updatedAt:
          type: string
      required:
      - name
      - updatedAt
      type: object
    ResourceType:
      enum:
      - workspace
      - target
      - build
      type: string
      x-enum-varnames:
      - ResourceTypeWorkspace
      - ResourceTypeTarget
      - ResourceTypeBuild
...
    Status:
      enum:
      - Unmodified
      - Untracked
      - Modified
      - Added
      - Deleted
      - Renamed
      - Copied
      - Updated but unmerged
      type: string
      x-enum-varnames:
      - Unmodified
      - Untracked
      - Modified
      - Added
      - Deleted
      - Renamed
      - Copied
      - UpdatedButUnmerged
    Target:
      example:
        default: true
        metadata:
          providerMetadata: providerMetadata
          targetId: targetId
          updatedAt: updatedAt
          uptime: 0
        targetConfig:
          deleted: true
          name: name
          options: options
          id: id
          providerInfo:
            agentlessTarget: true
            name: name
            runnerId: runnerId
            label: label
            version: version
        lastJob:
          createdAt: createdAt
          resourceId: resourceId
          action: null
          id: id
          state: null
          error: error
          resourceType: null
          updatedAt: updatedAt
        envVars:
          key: envVars
        name: name
        targetConfigId: targetConfigId
        id: id
        workspaces:
        - null
        - null
      properties:
        default:
          type: boolean
        envVars:
          additionalProperties:
            type: string
          type: object
        id:
          type: string
        lastJob:
          $ref: '#/components/schemas/Job'
        metadata:
          $ref: '#/components/schemas/TargetMetadata'
        name:
          type: string
        targetConfig:
          $ref: '#/components/schemas/TargetConfig'
        targetConfigId:
          type: string
        workspaces:
          items:
            $ref: '#/components/schemas/Workspace'
          type: array
      required:
      - default
      - envVars
      - id
      - name
      - targetConfig
      - targetConfigId
      - workspaces
      type: object
    TargetConfig:
      example:
        deleted: true
        name: name
        options: options
        id: id
        providerInfo:
          agentlessTarget: true
          name: name
          runnerId: runnerId
          label: label
          version: version
      properties:
        deleted:
          type: boolean
        id:
          type: string
        name:
          type: string
        options:
          description: JSON encoded map of options
          type: string
        providerInfo:
          $ref: '#/components/schemas/TargetProviderInfo'
      required:
      - deleted
      - id
      - name
      - options
      - providerInfo
      type: object
    TargetDTO:
      example:
        default: true
        metadata:
          providerMetadata: providerMetadata
          targetId: targetId
          updatedAt: updatedAt
          uptime: 0
        targetConfig:
          deleted: true
          name: name
          options: options
          id: id
          providerInfo:
            agentlessTarget: true
            name: name
            runnerId: runnerId
            label: label
            version: version
        lastJob:
          createdAt: createdAt
          resourceId: resourceId
          action: null
          id: id
          state: null
          error: error
          resourceType: null
          updatedAt: updatedAt
        envVars:
          key: envVars
        name: name
        targetConfigId: targetConfigId
        id: id
        state:
          name: null
          error: error
          updatedAt: updatedAt
        workspaces:
        - gitProviderConfigId: gitProviderConfigId
          image: image
          metadata:
            gitStatus:
              behind: 1
              fileStatus:
              - extra: extra
                name: name
                staging: null
                worktree: null
              - extra: extra
                name: name
                staging: null
                worktree: null
              ahead: 6
              branchPublished: true
              currentBranch: currentBranch
            updatedAt: updatedAt
            uptime: 5
            workspaceId: workspaceId
          providerMetadata: providerMetadata
          targetId: targetId
          envVars:
            key: envVars
          repository:
            owner: owner
            path: path
            name: name
            id: id
            source: source
            prNumber: 0
            branch: branch
            cloneTarget: null
            sha: sha
            url: url
          target:
            default: true
            metadata:
              providerMetadata: providerMetadata
              targetId: targetId
              updatedAt: updatedAt
              uptime: 0
            targetConfig:
              deleted: true
              name: name
              options: options
              id: id
              providerInfo:
                agentlessTarget: true
                name: name
                runnerId: runnerId
                label: label
                version: version
            lastJob:
              createdAt: createdAt
              resourceId: resourceId
              action: null
              id: id
              state: null
              error: error
              resourceType: null
              updatedAt: updatedAt
            envVars:
              key: envVars
            name: name
            targetConfigId: targetConfigId
            id: id
            workspaces:
            - null
            - null
          buildConfig:
            cachedBuild:
              image: image
              user: user
            devcontainer:
              filePath: filePath
          lastJob:
            createdAt: createdAt
            resourceId: resourceId
            action: null
            id: id
            state: null
            error: error
            resourceType: null
            updatedAt: updatedAt
          name: name
          id: id
          user: user
        - gitProviderConfigId: gitProviderConfigId
          image: image
          metadata:
            gitStatus:
              behind: 1
              fileStatus:
              - extra: extra
                name: name
                staging: null
                worktree: null
              - extra: extra
                name: name
                staging: null
                worktree: null
              ahead: 6
              branchPublished: true
              currentBranch: currentBranch
            updatedAt: updatedAt
            uptime: 5
            workspaceId: workspaceId
          providerMetadata: providerMetadata
          targetId: targetId
          envVars:
            key: envVars
          repository:
            owner: owner
            path: path
            name: name
            id: id
            source: source
            prNumber: 0
            branch: branch
            cloneTarget: null
            sha: sha
            url: url
          target:
            default: true
            metadata:
              providerMetadata: providerMetadata
              targetId: targetId
              updatedAt: updatedAt
              uptime: 0
            targetConfig:
              deleted: true
              name: name
              options: options
              id: id
              providerInfo:
                agentlessTarget: true
                name: name
                runnerId: runnerId
                label: label
                version: version
            lastJob:
              createdAt: createdAt
              resourceId: resourceId
              action: null
              id: id
              state: null
              error: error
              resourceType: null
              updatedAt: updatedAt
            envVars:
              key: envVars
            name: name
            targetConfigId: targetConfigId
            id: id
            workspaces:
            - null
            - null
          buildConfig:
            cachedBuild:
              image: image
              user: user
            devcontainer:
              filePath: filePath
          lastJob:
            createdAt: createdAt
            resourceId: resourceId
            action: null
            id: id
            state: null
            error: error
            resourceType: null
            updatedAt: updatedAt
          name: name
          id: id
          user: user
      properties:
        default:
          type: boolean
        envVars:
          additionalProperties:
            type: string
          type: object
        id:
          type: string
        lastJob:
          $ref: '#/components/schemas/Job'
        metadata:
          $ref: '#/components/schemas/TargetMetadata'
        name:
          type: string
        state:
          $ref: '#/components/schemas/ResourceState'
        targetConfig:
          $ref: '#/components/schemas/TargetConfig'
        targetConfigId:
          type: string
        workspaces:
          items:
            $ref: '#/components/schemas/Workspace'
          type: array
      required:
      - default
      - envVars
      - id
      - name
      - state
      - targetConfig
      - targetConfigId
      - workspaces
      type: object
    TargetMetadata:
      example:
        providerMetadata: providerMetadata
        targetId: targetId
        updatedAt: updatedAt
        uptime: 0
      properties:
        providerMetadata:
          type: string
        targetId:
          type: string
        updatedAt:
          type: string
        uptime:
          type: integer
      required:
      - targetId
      - updatedAt
      - uptime
      type: object
    TargetProviderInfo:
      example:
        agentlessTarget: true
        name: name
        runnerId: runnerId
        label: label
        version: version
      properties:
        agentlessTarget:
          type: boolean
        label:
          type: string
        name:
          type: string
        runnerId:
          type: string
        version:
          type: string
      required:
      - name
      - runnerId
      - version
      type: object
    Workspace:
      example:
        gitProviderConfigId: gitProviderConfigId
        image: image
        metadata:
          gitStatus:
            behind: 1
            fileStatus:
            - extra: extra
              name: name
              staging: null
              worktree: null
            - extra: extra
              name: name
              staging: null
              worktree: null
            ahead: 6
            branchPublished: true
            currentBranch: currentBranch
          updatedAt: updatedAt
          uptime: 5
          workspaceId: workspaceId
        providerMetadata: providerMetadata
        targetId: targetId
        envVars:
          key: envVars
        repository:
          owner: owner
          path: path
          name: name
          id: id
          source: source
          prNumber: 0
          branch: branch
          cloneTarget: null
          sha: sha
          url: url
        target:
          default: true
          metadata:
            providerMetadata: providerMetadata
            targetId: targetId
            updatedAt: updatedAt
            uptime: 0
          targetConfig:
            deleted: true
            name: name
            options: options
            id: id
            providerInfo:
              agentlessTarget: true
              name: name
              runnerId: runnerId
              label: label
              version: version
          lastJob:
            createdAt: createdAt
            resourceId: resourceId
            action: null
            id: id
            state: null
            error: error
            resourceType: null
            updatedAt: updatedAt
          envVars:
            key: envVars
          name: name
          targetConfigId: targetConfigId
          id: id
          workspaces:
          - null
          - null
        buildConfig:
          cachedBuild:
            image: image
            user: user
          devcontainer:
            filePath: filePath
        lastJob:
          createdAt: createdAt
          resourceId: resourceId
          action: null
          id: id
          state: null
          error: error
          resourceType: null
          updatedAt: updatedAt
        name: name
        id: id
        user: user
      properties:
        buildConfig:
          $ref: '#/components/schemas/BuildConfig'
        envVars:
          additionalProperties:
            type: string
          type: object
        gitProviderConfigId:
          type: string
        id:
          type: string
        image:
          type: string
        lastJob:
          $ref: '#/components/schemas/Job'
        metadata:
          $ref: '#/components/schemas/WorkspaceMetadata'
        name:
          type: string
        providerMetadata:
          type: string
        repository:
          $ref: '#/components/schemas/GitRepository'
        target:
          $ref: '#/components/schemas/Target'
        targetId:
          type: string
        user:
          type: string
      required:
      - envVars
      - id
      - image
      - name
      - repository
      - target
      - targetId
      - user
      type: object
...
    WorkspaceMetadata:
      example:
        gitStatus:
          behind: 1
          fileStatus:
          - extra: extra
            name: name
            staging: null
            worktree: null
          - extra: extra
            name: name
            staging: null
            worktree: null
          ahead: 6
          branchPublished: true
          currentBranch: currentBranch
        updatedAt: updatedAt
        uptime: 5
        workspaceId: workspaceId
      properties:
        gitStatus:
          $ref: '#/components/schemas/GitStatus'
        updatedAt:
          type: string
        uptime:
          type: integer
        workspaceId:
          type: string
      required:
      - gitStatus
      - updatedAt
      - uptime
      - workspaceId
      type: object
...
    models.JobAction:
      enum:
      - create
      - start
      - stop
      - restart
      - delete
      - force-delete
      - run
      type: string
      x-enum-varnames:
      - JobActionCreate
      - JobActionStart
      - JobActionStop
      - JobActionRestart
      - JobActionDelete
      - JobActionForceDelete
      - JobActionRun
    models.ResourceStateName:
      enum:
      - undefined
      - pending-run
      - running
      - run-successful
      - pending-create
      - creating
      - pending-start
      - starting
      - started
      - pending-stop
      - stopping
      - stopped
      - pending-restart
      - error
      - unresponsive
      - pending-delete
      - pending-forced-delete
      - deleting
      - deleted
      type: string
      x-enum-varnames:
      - ResourceStateNameUndefined
      - ResourceStateNamePendingRun
      - ResourceStateNameRunning
      - ResourceStateNameRunSuccessful
      - ResourceStateNamePendingCreate
      - ResourceStateNameCreating
      - ResourceStateNamePendingStart
      - ResourceStateNameStarting
      - ResourceStateNameStarted
      - ResourceStateNamePendingStop
      - ResourceStateNameStopping
      - ResourceStateNameStopped
      - ResourceStateNamePendingRestart
      - ResourceStateNameError
      - ResourceStateNameUnresponsive
      - ResourceStateNamePendingDelete
      - ResourceStateNamePendingForcedDelete
      - ResourceStateNameDeleting
      - ResourceStateNameDeleted
...
  securitySchemes:
    Bearer:
      description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
      in: header
      name: Authorization
      type: apiKey
...